선영이꺼 합병 후 충돌 수정 + 앱탈퇴기능 + LoginActivity/SuccessLogin 수정
