+++ app/src/main/res/layout/activity_success_login.xml
@@ -5,13 +5,35 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:gravity="center"
+    android:orientation="vertical"
     tools:context=".SuccessLogin">
 
 
+    <TextView
+        android:id="@+id/textView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:text="로그인 성공했을 때 화면" />
+
     <Button
         android:id="@+id/kakao_logout_btn"
-        android:layout_width="wrap_content"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_weight="1"
         android:text="카카오 로그아웃" />
+
+    <Button
+        android:id="@+id/delete_session_btn"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:text="앱 탈퇴하기" />
+
+    <Button
+        android:id="@+id/redirect_login_btn"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_weight="1"
+        android:text="로그인 액티비티로 돌아가기" />
 </LinearLayout>
